src/App/requirements.txt
fastapi
uvicorn[standard]
# uvicorn removed and added above to allow websocket support
jinja2
azure-identity
python-dotenv
python-multipart
httpx